[lexical-table] Refactor: Delegate insertRawText to $insertTableIntoGrid
Move $insertTableIntoGrid from LexicalTablePluginHelpers to
LexicalTableUtils to eliminate a circular dependency (Selection ↔
PluginHelpers). Refactor insertRawText from 80-line direct
implementation to build a temporary TableNode and delegate to
$insertTableIntoGrid, reusing existing row/column expansion and
merged cell unmerging logic.
