+++ packages/lexical-table/src/LexicalTablePluginHelpers.ts
@@ -67,15 +67,11 @@ import {
   registerTableWindowHandlers,
 } from './LexicalTableSelectionHelpers';
 import {
-  $computeTableCellRectBoundary,
   $computeTableMap,
   $computeTableMapSkipCellCheck,
   $createTableNodeWithDimensions,
   $getNodeTriplet,
-  $insertTableColumnAtNode,
-  $insertTableRowAtNode,
-  $mergeCells,
-  $unmergeCellNode,
+  $insertTableIntoGrid,
 } from './LexicalTableUtils';
 
 function $insertTable(
@@ -489,183 +485,6 @@ function $tableSelectionInsertClipboardNodesCommand(
   }
 
   // If we reached this point, there's a table in the selection and nested tables are not allowed - reject the paste.
-  return true;
-}
-
-function $insertTableIntoGrid(
-  tableNode: TableNode,
-  selection: RangeSelection | TableSelection,
-) {
-  const anchorAndFocus = selection.getStartEndPoints();
-  const isTableSelection = $isTableSelection(selection);
-
-  if (anchorAndFocus === null) {
-    return false;
-  }
-
-  const [anchor, focus] = anchorAndFocus;
-  const [anchorCellNode, anchorRowNode, gridNode] = $getNodeTriplet(anchor);
-  const focusCellNode = $findMatchingParent(focus.getNode(), n =>
-    $isTableCellNode(n),
-  );
-
-  if (
-    !$isTableCellNode(anchorCellNode) ||
-    !$isTableCellNode(focusCellNode) ||
-    !$isTableRowNode(anchorRowNode) ||
-    !$isTableNode(gridNode)
-  ) {
-    return false;
-  }
-
-  const [initialGridMap, anchorCellMap, focusCellMap] = $computeTableMap(
-    gridNode,
-    anchorCellNode,
-    focusCellNode,
-  );
-  const [templateGridMap] = $computeTableMapSkipCellCheck(
-    tableNode,
-    null,
-    null,
-  );
-  const initialRowCount = initialGridMap.length;
-  const initialColCount = initialRowCount > 0 ? initialGridMap[0].length : 0;
-
-  // If we have a range selection, we'll fit the template grid into the
-  // table, growing the table if necessary.
-  let startRow = anchorCellMap.startRow;
-  let startCol = anchorCellMap.startColumn;
-  let affectedRowCount = templateGridMap.length;
-  let affectedColCount = affectedRowCount > 0 ? templateGridMap[0].length : 0;
-
-  if (isTableSelection) {
-    // If we have a table selection, we'll only modify the cells within
-    // the selection boundary.
-    const selectionBoundary = $computeTableCellRectBoundary(
-      initialGridMap,
-      anchorCellMap,
-      focusCellMap,
-    );
-    const selectionRowCount =
-      selectionBoundary.maxRow - selectionBoundary.minRow + 1;
-    const selectionColCount =
-      selectionBoundary.maxColumn - selectionBoundary.minColumn + 1;
-    startRow = selectionBoundary.minRow;
-    startCol = selectionBoundary.minColumn;
-    affectedRowCount = Math.min(affectedRowCount, selectionRowCount);
-    affectedColCount = Math.min(affectedColCount, selectionColCount);
-  }
-
-  // Step 1: Unmerge all merged cells within the affected area
-  let didPerformMergeOperations = false;
-  const lastRowForUnmerge =
-    Math.min(initialRowCount, startRow + affectedRowCount) - 1;
-  const lastColForUnmerge =
-    Math.min(initialColCount, startCol + affectedColCount) - 1;
-  const unmergedKeys = new Set<NodeKey>();
-  for (let row = startRow; row <= lastRowForUnmerge; row++) {
-    for (let col = startCol; col <= lastColForUnmerge; col++) {
-      const cellMap = initialGridMap[row][col];
-      if (unmergedKeys.has(cellMap.cell.getKey())) {
-        continue; // cell was a merged cell that was already handled
-      }
-      if (cellMap.cell.__rowSpan === 1 && cellMap.cell.__colSpan === 1) {
-        continue; // cell is not a merged cell
-      }
-      $unmergeCellNode(cellMap.cell);
-      unmergedKeys.add(cellMap.cell.getKey());
-      didPerformMergeOperations = true;
-    }
-  }
-
-  let [interimGridMap] = $computeTableMapSkipCellCheck(
-    gridNode.getWritable(),
-    null,
-    null,
-  );
-
-  // Step 2: Expand current table (if needed)
-  const rowsToInsert = affectedRowCount - initialRowCount + startRow;
-  for (let i = 0; i < rowsToInsert; i++) {
-    const cellMap = interimGridMap[initialRowCount - 1][0];
-    $insertTableRowAtNode(cellMap.cell);
-  }
-  const colsToInsert = affectedColCount - initialColCount + startCol;
-  for (let i = 0; i < colsToInsert; i++) {
-    const cellMap = interimGridMap[0][initialColCount - 1];
-    $insertTableColumnAtNode(cellMap.cell, true, false);
-  }
-
-  [interimGridMap] = $computeTableMapSkipCellCheck(
-    gridNode.getWritable(),
-    null,
-    null,
-  );
-
-  // Step 3: Merge cells and set cell content, to match template grid
-  for (let row = startRow; row < startRow + affectedRowCount; row++) {
-    for (let col = startCol; col < startCol + affectedColCount; col++) {
-      const templateRow = row - startRow;
-      const templateCol = col - startCol;
-      const templateCellMap = templateGridMap[templateRow][templateCol];
-      if (
-        templateCellMap.startRow !== templateRow ||
-        templateCellMap.startColumn !== templateCol
-      ) {
-        continue; // cell is a merged cell that was already handled
-      }
-
-      const templateCell = templateCellMap.cell;
-      if (templateCell.__rowSpan !== 1 || templateCell.__colSpan !== 1) {
-        const cellsToMerge = [];
-        const lastRowForMerge =
-          Math.min(row + templateCell.__rowSpan, startRow + affectedRowCount) -
-          1;
-        const lastColForMerge =
-          Math.min(col + templateCell.__colSpan, startCol + affectedColCount) -
-          1;
-        for (let r = row; r <= lastRowForMerge; r++) {
-          for (let c = col; c <= lastColForMerge; c++) {
-            const cellMap = interimGridMap[r][c];
-            cellsToMerge.push(cellMap.cell);
-          }
-        }
-        $mergeCells(cellsToMerge);
-        didPerformMergeOperations = true;
-      }
-
-      const {cell} = interimGridMap[row][col];
-      const backgroundColor = templateCell.getBackgroundColor();
-      if (backgroundColor !== null && backgroundColor !== undefined) {
-        cell.setBackgroundColor(backgroundColor);
-      }
-      const originalChildren = cell.getChildren();
-      templateCell.getChildren().forEach(child => {
-        if ($isTextNode(child)) {
-          const paragraphNode = $createParagraphNode();
-          paragraphNode.append(child);
-          cell.append(child);
-        } else {
-          cell.append(child);
-        }
-      });
-      originalChildren.forEach(n => n.remove());
-    }
-  }
-
-  if (isTableSelection && didPerformMergeOperations) {
-    // reset the table selection in case the anchor or focus cell was
-    // removed via merge operations
-    const [finalGridMap] = $computeTableMapSkipCellCheck(
-      gridNode.getWritable(),
-      null,
-      null,
-    );
-    const newAnchorCellMap =
-      finalGridMap[anchorCellMap.startRow][anchorCellMap.startColumn];
-    newAnchorCellMap.cell.selectEnd();
-  }
-
   return true;
 }
