+++ packages/lexical-table/src/__tests__/unit/LexicalTableSelection.test.tsx
@@ -232,7 +232,7 @@ describe('table selection', () => {
         );
       });
 
-      test('skips already-filled merged cells', () => {
+      test('unmerges and fills merged cells', () => {
         testEnv.editor.update(
           () => {
             // Build a 3x3 table where cell (0,0) spans 2 cols
@@ -270,10 +270,8 @@ describe('table selection', () => {
               .getChildren()
               .filter($isTableRowNode)
               .map(row => row.getChildren().map(cell => cell.getTextContent()));
-            // topLeft spans cols 0-1, TSV cell (0,0) fills it with "A",
-            // TSV cell (0,1) maps to the same merged node — skipped
-            expect(texts[0][0]).toBe('A');
-            expect(texts[0][1]).toBe('C');
+            // Merged cell is unmerged, each cell gets its TSV value
+            expect(texts[0]).toEqual(['A', 'B', 'C']);
             expect(texts[1]).toEqual(['D', 'E', 'F']);
           },
           {discrete: true},
